moved acceptance tests
